fix comment 8
