+++ fe/fe-core/src/main/java/org/apache/doris/analysis/BaseTableRef.java
@@ -58,8 +58,6 @@ public class BaseTableRef extends TableRef {
     public TupleDescriptor createTupleDescriptor(Analyzer analyzer) {
         TupleDescriptor result = analyzer.getDescTbl().createTupleDescriptor();
         result.setTable(table);
-        result.setSampleTabletIds(sampleTabletIds);
-        result.setTableSample(tableSample);
         return result;
     }
